+++ biom/_subsample.pyx
@@ -11,7 +11,7 @@ import numpy as np
 cimport numpy as cnp
 
 
-def _subsample(arr, n, with_replacement):
+def _subsample(arr, n, with_replacement, rng):
     """Subsample non-zero values of a sparse array
 
     Parameters
@@ -20,7 +20,12 @@ def _subsample(arr, n, with_replacement):
         A 1xM sparse vector
     n : int
         Number of items to subsample from `arr`
-    
+    with_replacement : bool
+        Whether to permute or use multinomial sampling
+    rng : Generator instance
+        A random generator. This will likely be an instance returned 
+        by np.random.default_rng
+
     Returns
     -------
     ndarray
@@ -34,7 +39,7 @@ def _subsample(arr, n, with_replacement):
     cdef:
         cnp.int64_t counts_sum
         cnp.ndarray[cnp.float64_t, ndim=1] data = arr.data
-        cnp.ndarray[cnp.int32_t, ndim=1] data_i = arr.data.astype(np.int32)
+        cnp.ndarray[cnp.int64_t, ndim=1] data_i = arr.data.astype(np.int64)
         cnp.ndarray[cnp.float64_t, ndim=1] result
         cnp.ndarray[cnp.int32_t, ndim=1] indices = arr.indices
         cnp.ndarray[cnp.int32_t, ndim=1] indptr = arr.indptr
@@ -49,7 +54,7 @@ def _subsample(arr, n, with_replacement):
         
         if with_replacement:
             pvals = data[start:end] / counts_sum
-            data[start:end] = np.random.multinomial(n, pvals)
+            data[start:end] = rng.multinomial(n, pvals)
         else:
             if counts_sum < n:
                 data[start:end] = 0
@@ -57,7 +62,7 @@ def _subsample(arr, n, with_replacement):
 
             r = np.arange(length, dtype=np.int32)
             unpacked = np.repeat(r, data_i[start:end])
-            permuted = np.random.permutation(unpacked)[:n]
+            permuted = rng.permutation(unpacked)[:n]
 
             result = np.zeros(length, dtype=np.float64)
             for idx in range(permuted.shape[0]):